RECservice/requirements.txt
pip>=21.3.1
setuptools>=60.2.0
wheel>=0.37.1
sentence-transformers
numpy
faiss-cpu
mysql-connector-python
python-dotenv>=0.20.0
pandas
flask
flask_api
waitress
flask_cors